PD-10: Fixed Optional methods, changed double to BigDecimal and initialize method name.

+++ src/pd10/User.java
@@ -3,10 +3,12 @@ package pd10;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor(staticName = "of")
 public final class User {
     private final String name;
     private final TransportType preferredTransport;
-    private final Double budget;
+    private final BigDecimal budget;
 }
